Extract draft audio response helper
